Update syllabus, schedule, lab 1. Add lecture notes 4.
Add lecture notes 4 to the repository. Push back lab 1 to Wednesday, so two days back. Update lab1.rst and the syllabus to match. Note that the remainder of labs have a due date of TBA due to this change. Add the Canvas link to the syllabus. Update the schedule to account for the change.